requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.2
requests==2.31.0
openai>=1.0.0
plotly==5.15.0
python-dotenv==1.0.0